fixed routes typo
