+++ laughfactory/index.py
@@ -1,0 +1,7 @@
+from flask import ( Blueprint, render_template, request, redirect ) 
+
+bp = Blueprint('index', __name__, url_prefix="/index")
+
+@bp.route('/')
+def index():
+    return render_template('index.html')
