+++ laughfactory/login.py
@@ -1,0 +1,7 @@
+from flask import ( Blueprint, render_template, request, redirect ) 
+
+bp = Blueprint('login', __name__, url_prefix="/login")
+
+@bp.route('/')
+def login():
+    return render_template('login.html')
