--- laughfactory/meme.py
@@ -1,4 +1,0 @@
-from flask import ( Blueprint, render_template, request, redirect ) 
-
-bp = Blueprint('meme', __name__, url_prefix="/meme")
-
