+++ laughfactory/models.py
@@ -1,3 +1,5 @@
+# ignore this code for now. wait until database tested
+
 from flask_sqlalchemy import SQLAlchemy 
 
 db = SQLAlchemy()
