+++ laughfactory/templates/upload.html
@@ -7,6 +7,6 @@
     <title>Document</title>
 </head>
 <body>
-    <h1>HIIIIII</h1>
+    <h1>This is upload page</h1>
 </body>
 </html>
